for functions java I made another word to print four times, for return java I made it square a number and add it to another squared number, for robots java I made it do another three things.

+++ app/src/main/java/_training/Return.java
@@ -1,7 +1,7 @@
 package _training;
 
 public class Return {
-    static float square(float num){
+    static int square(int num){
         return num*num;
     }
     public static void main(String[] args){
